pyproject.toml
[project]
name = "final-year"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.5",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.38.0",
]
